+++ fe/lib/components/JobCategorySuggest/index.ts
@@ -1,0 +1,1 @@
+export { JobCategorySuggest } from './JobCategorySuggest';