+++ fe/lib/components/index.ts
@@ -1,3 +1,4 @@
 export { Example } from './Example/Example';
-export { LocationSuggest } from './LocationSuggest/LocationSuggest';
 export { JobCategorySelect } from './JobCategorySelect/JobCategorySelect';
+export { JobCategorySuggest } from './JobCategorySuggest/JobCategorySuggest';
+export { LocationSuggest } from './LocationSuggest/LocationSuggest';