+++ fe/lib/utils/flattenResourceByKey.ts
@@ -1,0 +1,7 @@
+export const flattenResourceByKey = <T>(resource: T, key: string): T[] => {
+  const castedResource = resource as any;
+
+  return [resource].concat(
+    castedResource[key] ? flattenResourceByKey(castedResource[key], key) : [],
+  );
+};